+++ test/legacy_test/test_detection.py
@@ -16,7 +16,6 @@ import contextlib
 import unittest
 
 import numpy as np
-from test_imperative_base import new_program_scope
 
 import paddle
 import paddle.fluid as fluid
@@ -25,6 +24,17 @@ from paddle.fluid.dygraph import base
 from paddle.fluid.framework import Program, program_guard
 
 paddle.enable_static()
+
+
+@contextlib.contextmanager
+def new_program_scope(main=None, startup=None, scope=None):
+    prog = main if main else fluid.Program()
+    startup_prog = startup if startup else fluid.Program()
+    scope = scope if scope else fluid.core.Scope()
+    with fluid.scope_guard(scope):
+        with fluid.program_guard(prog, startup_prog):
+            with fluid.unique_name.guard():
+                yield
 
 
 class LayerTest(unittest.TestCase):
